# my-blog-end/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/article/add": {
            "post": {
                "description": "发博客",
                "tags": [
                    "article"
                ],
                "summary": "发博客",
                "parameters": [
                    {
                        "description": "ReqArticle",
                        "name": "ReqArticle",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ReqArticle"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/article/getSomeoneArticle": {
            "post": {
                "description": "某人的博客列表，不传userId查询当前用户的",
                "tags": [
                    "article"
                ],
                "summary": "某人的博客列表",
                "parameters": [
                    {
                        "description": "QueryArticleBody",
                        "name": "QueryArticleBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.QueryArticleBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.PageResult-controller_ArticleVo"
                        }
                    }
                }
            }
        },
...
        "/auth/login": {
            "post": {
                "description": "用户登录",
                "tags": [
                    "auth"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "description": "ReqUser",
                        "name": "ReqUser",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ReqUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "用户注册",
                "tags": [
                    "auth"
                ],
                "summary": "注册",
                "parameters": [
                    {
                        "description": "ReqUser",
                        "name": "ReqUser",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ReqUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/common/article/getById/:ID": {
            "get": {
                "description": "博客详情",
                "tags": [
                    "article"
                ],
                "summary": "博客详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ArticleVo"
                        }
                    }
                }
            }
        },
        "/common/article/query": {
            "post": {
                "description": "博客列表",
                "tags": [
                    "article"
                ],
                "summary": "博客列表",
                "parameters": [
                    {
                        "description": "QueryArticleBody",
                        "name": "QueryArticleBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.QueryArticleBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.PageResult-controller_ArticleVo"
                        }
                    }
                }
            }
        },
...
        "/delete/:ID": {
            "delete": {
                "description": "删除博客(当前用户的)",
                "tags": [
                    "article"
                ],
                "summary": "删除博客",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/like/add": {
            "post": {
                "description": "点赞",
                "tags": [
                    "like"
                ],
                "summary": "点赞",
                "parameters": [
                    {
                        "description": "ReqLike",
                        "name": "ReqLike",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ReqLike"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/like/cancel": {
            "post": {
                "description": "取消点赞",
                "tags": [
                    "like"
                ],
                "summary": "取消点赞",
                "parameters": [
                    {
                        "description": "ReqLike",
                        "name": "ReqLike",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ReqLike"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.ArticleVo": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "integer"
                },
                "authId": {
                    "type": "integer"
                },
                "authName": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "createdTime": {
                    "type": "string"
                },
                "likeFlag": {
                    "type": "boolean"
                },
                "likeNum": {
                    "type": "integer"
                },
                "parentId": {
                    "type": "integer"
                },
                "replyNum": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "controller.PageResult-controller_ArticleVo": {
            "type": "object",
            "properties": {
                "all": {
                    "type": "integer"
                },
                "content": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.ArticleVo"
                    }
                },
                "page_num": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                }
            }
        },
...
        "controller.QueryArticleBody": {
            "type": "object",
            "properties": {
                "descCreatedTime": {
                    "type": "boolean"
                },
                "key": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
        "controller.ReqArticle": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "controller.ReqLike": {
            "type": "object",
            "required": [
                "articleId"
            ],
            "properties": {
                "articleId": {
                    "type": "integer"
                }
            }
        },
...
        "controller.ReqUser": {
            "type": "object",
            "required": [
                "name",
                "password"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
    }
}